convert e2e test to unit test
